Add partial to customize og:image meta property
Adds a new partial — `head/og-image.html` — that will render the value to be used
for `og:image` meta tags. The changes preserve the existing Hugo OpenGraph template's
behavior of checking and using the frontmatter or specific format of image names
as the `og:image` but if that doesn't work, it will delegate to the default partial which
uses `assets/images/og-image.{ext}` as the image.

This partial can be overriden to allow dynamic `og:image` URLs via a third-party service.

`_internal/opengraph.html` source is copied from Hugo to customize and support
this feature. Docs are updated to mention its usage.